requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-environ==0.11.2
djangorestframework==3.15.1
idna==3.7
Pillow==10.1.0
psycopg2-binary==2.9.9
python-slugify==8.0.4
requests==2.31.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
